Enhance linkedin-cleanup.js
Linkedin also showed ads for selling Premium. I don't want that so I enhanced the script to also hide that.

I made it a bit more flexible.

You can now hide anything by passing a selector for the object to hide and a regexp for the span-text it has to contain.